don't just merge my pull request with your changes!
changes:
generating shaders should not be a part of the build step, we don't
generate shaders enough to make it the wasted time worth. It should just
be a separate step. Also, if we don't commit the shaders, then the
project would be in a unbuildable state. The shaders only need to be
rebuild when we change them.

Also the Tools in the tools directory are specific to this project,
there is no need to make them work for other projects.

Also why tf did you add a gitkeep for the shaders directory if we so or
so create it in the python script.